Suppliers Work 100%

+++ database/migrations/2025_03_02_130935_add_status_to_debts_table.php
@@ -3,17 +3,19 @@
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Support\Facades\DB;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::table('debts', function (Blueprint $table) {
-            $table->string('status')->default('unpaid')->after('amount');
-        });
+        if (!Schema::hasColumn('debts', 'status')) {
+            Schema::table('debts', function (Blueprint $table) {
+                $table->string('status')->default('unpaid')->after('amount');
+            });
+        }
     }
 
     /**
@@ -21,8 +23,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('debts', function (Blueprint $table) {
-            $table->dropColumn('status');
-        });
+        if (Schema::hasColumn('debts', 'status')) {
+            Schema::table('debts', function (Blueprint $table) {
+                $table->dropColumn('status');
+            });
+        }
     }
 };
